FaceID_WebApp/requirements.txt
streamlit>=1.28.0
tensorflow>=2.10.0,<2.16.0
opencv-python>=4.5.0
mtcnn>=0.1.0
numpy>=1.21.0,<2.0.0
Pillow>=8.0.0
matplotlib>=3.3.0
pandas>=1.3.0
scikit-learn>=1.0.0
protobuf>=3.19.0,<4.0.0 